requirements.txt
langgraph>=0.0.10
langchain-anthropic>=0.1.1
langchain-core>=0.1.27
PyGithub>=2.2.0
GitPython>=3.1.42
pydantic>=2.6.1
python-dotenv>=1.0.0
